Corrigir classe CalculadoraMediaNumerosIndeterminados, colocando o resultado da operação em tipo double, com o intuito de torná-lo mais preciso.

+++ src/exerciciosEstruturaRepeticao/exercicio2/CalculadoraMediaNumerosIndeterminados.java
@@ -21,7 +21,7 @@ public class CalculadoraMediaNumerosIndeterminados {
                 } else {
                     soma += numero;
                 }
-                int media = soma / quantidadeNumeros;
+                double media = (double) soma / (double) quantidadeNumeros;
                 System.out.println("A média dos números informados é: " + media);
             }
             input.close();
